use RootState throughout
