+++ web/src/components/ExportMenuItem/ExportMenuItem.connector.ts
@@ -1,10 +1,8 @@
 import { connect } from 'react-redux'
+import { RootState } from '../../redux/reducer'
 import ExportMenuItem from './ExportMenuItem.component'
 
-function mapDispatchToProps(dispatch) {
-  return {}
-}
-function mapStateToProps(state) {
+function mapStateToProps(state: RootState) {
   const {
     ui: { activeProject },
   } = state
@@ -15,8 +13,8 @@ function mapStateToProps(state) {
   const outcomeComments = state.projects.outcomeComments[activeProject] || {}
   const outcomeVotes = state.projects.outcomeVotes[activeProject] || {}
   const entryPoints = state.projects.entryPoints[activeProject] || {}
-  const activeProjectMeta = state.projects.projectMeta[activeProject] || {}
-  const projectName = activeProjectMeta.name || ''
+  const activeProjectMeta = state.projects.projectMeta[activeProject] || {name: ''}
+  const projectName = activeProjectMeta.name
 
   return {
     projectName,
@@ -31,6 +29,10 @@ function mapStateToProps(state) {
       entryPoints,
     },
   }
+}
+
+function mapDispatchToProps(dispatch) {
+  return {}
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(ExportMenuItem)
