+++ web/src/components/PeoplePicker/PeoplePicker.connector.ts
@@ -1,15 +1,17 @@
 import { connect } from 'react-redux'
+import moment from 'moment'
 import {
   createOutcomeMember,
   deleteOutcomeMember,
 } from '../../redux/persistent/projects/outcome-members/actions'
-import moment from 'moment'
-import PeoplePicker from './PeoplePicker.component'
+import { AgentPubKeyB64, HeaderHashB64 } from '../../types/shared'
 import ProjectsZomeApi from '../../api/projectsApi'
 import { getAppWs } from '../../hcWebsockets'
 import { cellIdFromString } from '../../utils'
+import { RootState } from '../../redux/reducer'
+import PeoplePicker from './PeoplePicker.component'
 
-function mapStateToProps(state, ownProps) {
+function mapStateToProps(state: RootState, ownProps) {
   const outcomeHeaderHash = state.ui.outcomeForm.isOpen
     ? state.ui.outcomeForm.editAddress
     : state.ui.expandedView.outcomeHeaderHash
@@ -18,17 +20,20 @@ function mapStateToProps(state, ownProps) {
   const members = state.projects.members[projectId] || {}
   const membersOfOutcome = Object.keys(outcomeMembers)
     .map((address) => outcomeMembers[address])
-    .filter((outcomeMember) => outcomeMember.outcomeHeaderHash === outcomeHeaderHash)
+    .filter(
+      (outcomeMember) => outcomeMember.outcomeHeaderHash === outcomeHeaderHash
+    )
   // just in case we've received a 'member' before the agents profile
   // filter out any missing profiles for now
-  const agents = Object.keys(members).map((address) => state.agents[address]).filter((agent) => agent)
+  const agents = Object.keys(members)
+    .map((address) => state.agents[address])
+    .filter((agent) => agent)
   return {
     agentAddress: state.agentAddress,
     people: agents.map((agent) => {
       const member = membersOfOutcome.find(
         (outcomeMember) => outcomeMember.agentAddress === agent.address
       )
-        console.log(member)
       return {
         ...agent, // address, name, avatarUrl
         is_member: member ? true : false,
@@ -43,25 +48,27 @@ function mapDispatchToProps(dispatch, ownProps) {
   const { projectId: cellIdString } = ownProps
   const cellId = cellIdFromString(cellIdString)
   return {
-    createOutcomeMember: async (outcomeHeaderHash, agentAddress, creatorAgentPubKey) => {
+    createOutcomeMember: async (
+      outcomeHeaderHash: HeaderHashB64,
+      memberAgentPubKey: AgentPubKeyB64,
+      creatorAgentPubKey: AgentPubKeyB64
+    ) => {
       const appWebsocket = await getAppWs()
       const projectsZomeApi = new ProjectsZomeApi(appWebsocket)
       const outcomeMember = await projectsZomeApi.outcomeMember.create(cellId, {
         outcomeHeaderHash,
-        memberAgentPubKey: agentAddress,
+        memberAgentPubKey: memberAgentPubKey,
         creatorAgentPubKey,
         unixTimestamp: moment().unix(),
-        isImported: false
+        isImported: false,
       })
-      return dispatch(
-        createOutcomeMember(cellIdString, outcomeMember)
-        )
-      },
-    deleteOutcomeMember: async (payload) => {
+      return dispatch(createOutcomeMember(cellIdString, outcomeMember))
+    },
+    deleteOutcomeMember: async (headerHash: HeaderHashB64) => {
       const appWebsocket = await getAppWs()
       const projectsZomeApi = new ProjectsZomeApi(appWebsocket)
-      const deleteOutcomeMemberHash = await projectsZomeApi.outcomeMember.delete(cellId, payload)
-      return dispatch(deleteOutcomeMember(cellIdString, deleteOutcomeMemberHash))
+      await projectsZomeApi.outcomeMember.delete(cellId, headerHash)
+      return dispatch(deleteOutcomeMember(cellIdString, headerHash))
     },
   }
 }
