+++ web/src/components/ProjectSettingsModal/ProjectSettingsModal.connector.ts
@@ -1,12 +1,14 @@
 import { connect } from 'react-redux'
-import './ProjectSettingsModal.scss'
 import { updateProjectMeta } from '../../redux/persistent/projects/project-meta/actions'
-import ProjectSettingsModal from './ProjectSettingsModal.component'
 import ProjectsZomeApi from '../../api/projectsApi'
 import { getAppWs } from '../../hcWebsockets'
 import { cellIdFromString } from '../../utils'
+import ProjectSettingsModal from './ProjectSettingsModal.component'
+import { RootState } from '../../redux/reducer'
+import { CellIdString, HeaderHashB64 } from '../../types/shared'
+import { ProjectMeta } from '../../types'
 
-function mapStateToProps(_state) {
+function mapStateToProps(_state: RootState) {
   // props for the componen
   return {}
 }
@@ -14,16 +16,20 @@ function mapStateToProps(_state) {
 function mapDispatchToProps(dispatch) {
   // props for the component
   return {
-    updateProjectMeta: async (entry, headerHash, cellIdString) => {
+    updateProjectMeta: async (
+      projectMeta: ProjectMeta,
+      headerHash: HeaderHashB64,
+      cellIdString: CellIdString
+    ) => {
       const appWebsocket = await getAppWs()
       const projectsZomeApi = new ProjectsZomeApi(appWebsocket)
       const cellId = cellIdFromString(cellIdString)
-      // @ts-ignore
-      const updatedProjectMeta = await projectsZomeApi.projectMeta.update(cellId, { entry, headerHash })
-      return dispatch(
-        updateProjectMeta(cellIdString, updatedProjectMeta)
+      const updatedProjectMeta = await projectsZomeApi.projectMeta.update(
+        cellId,
+        { entry: projectMeta, headerHash }
       )
-    }
+      return dispatch(updateProjectMeta(cellIdString, updatedProjectMeta))
+    },
   }
 }
